Fix: Standardize and correct typos in component and folder names (Accomplishments, BackgroundAnimation) and update all references
