+++ src/components/BackgroundAnimation/BackgroundAnimation.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const BackgroundAnimation = () => (
+  <div>
+    <svg
+      className="BgAnimation__svg"
+      viewBox="0 0 602 602"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      {/* ...existing SVG code... */}
+    </svg>
+  </div>
+);
+
+export default BackgroundAnimation;
